feat: update for removal of readdc.com
I have to parse the comoxilogy website now. There is no way to search by
publisher, so for each book that is found, I need to go to its page and
check the publisher before adding it to the list.

The publisher was not populated with the requests library, so I had to
use requests-html to render the JavaScript.